Clean up and organize utils
