+++ src/commands/integrations/init/yaml/fixtures/specs/simple-integration.yaml
