+++ src/index.ts
@@ -73,7 +73,7 @@ import InstancesFlowConfigsTestCommand from "./commands/instances/flow-configs/t
 import IntegrationsFlowsListCommand from "./commands/integrations/flows/list.js";
 import IntegrationsFlowsTestCommand from "./commands/integrations/flows/test.js";
 import IntegrationsInitFlowCommand from "./commands/integrations/init/flow.js";
-import IntegrationsInitFromYamlCommand from "./commands/integrations/init/yaml.js";
+import IntegrationsInitFromYamlCommand from "./commands/integrations/init/yaml/yaml.js";
 import IntegrationsInitCommand from "./commands/integrations/init/index.js";
 import IntegrationsInitIntegrationCommand from "./commands/integrations/init/integration.js";
 import IntegrationsVersionsCommand from "./commands/integrations/versions/index.js";
